pagesのディレクトリを追加
